+++ .templates/new-module/tests/react/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import [-replace-name-capital-] from './[-replace-name-].js';
+
+
+ReactDOM.render(
+	<div>
+		<[-replace-name-capital-]>
+		</[-replace-name-capital-]>
+	</div>,
+
+	document.getElementById('root'),
+);